refactor(mcp): extract proxy tools to tools_proxy.py
